Extract out BacklogViewSwitcher component
