+++ frontend/src/components/Backlog/Backlog.jsx
@@ -1,10 +1,9 @@
-// @flow
+// @flow strict
 
 import * as React from 'react';
-import { Button } from 'semantic-ui-react';
 import type { BacklogDisplayOption } from './backlog-types';
-import styles from './Backlog.css';
 import BacklogDaysContainer from './BacklogDaysContainer';
+import BacklogViewSwitcher from './BacklogViewSwitcher';
 
 type ComponentState = {| displayOption: BacklogDisplayOption |};
 
@@ -19,24 +18,9 @@ export default class Backlog extends React.Component<{}, ComponentState> {
 
   render() {
     const { displayOption } = this.state;
-    const setDisplayOption = (newDisplayOption: BacklogDisplayOption) => (
-      () => this.setState({ displayOption: newDisplayOption })
-    );
-    const switchComponent = (
-      <div className={styles.BacklogControl}>
-        <span className={styles.BacklogControlPadding} />
-        <Button.Group>
-          <Button onClick={setDisplayOption('FOUR_DAYS')}>4D</Button>
-          <Button.Or />
-          <Button onClick={setDisplayOption('BIWEEKLY')}>2W</Button>
-          <Button.Or />
-          <Button onClick={setDisplayOption('MONTHLY')}>M</Button>
-        </Button.Group>
-      </div>
-    );
     return (
       <div>
-        {switchComponent}
+        <BacklogViewSwitcher onChange={o => this.setState({ displayOption: o })} />
         <BacklogDaysContainer displayOption={displayOption} />
       </div>
     );
